Keep PR automation Pi-only
