--- .github/workflows/llm-pr-review.yml
@@ -1,108 +1,0 @@
-name: LLM PR Review
-
-"on":
-  pull_request_target:
-    types: [opened, reopened, synchronize, ready_for_review]
-  workflow_dispatch:
-    inputs:
-      pr_number:
-        description: Open pull request number to review
-        required: true
-        type: string
-
-permissions:
-  contents: read
-  pull-requests: write
-
-concurrency:
-  group: llm-pr-review-${{ github.event_name }}-${{ github.event.pull_request.number || inputs.pr_number }}
-  cancel-in-progress: true
-
-jobs:
-  review:
-    name: Hosted review
-    if: >-
-      ${{ github.event_name == 'workflow_dispatch' ||
-      !github.event.pull_request.draft }}
-    runs-on: ubuntu-latest
-    environment: llm-pr-review
-    timeout-minutes: 20
-    steps:
-      - name: Resolve review target
-        id: target
-        env:
-          EVENT_NAME: ${{ github.event_name }}
-          EVENT_PATH: ${{ github.event_path }}
-          INPUT_PR_NUMBER: ${{ inputs.pr_number }}
-          REPOSITORY: ${{ github.repository }}
-          GH_TOKEN: ${{ github.token }}
-          TEMPORARY_DIRECTORY: ${{ runner.temp }}
-        run: |
-          set -euo pipefail
-          python3 - <<'PY'
-          import json
-          import os
-          import re
-          from pathlib import Path
-          from urllib.request import Request, urlopen
-
-          if os.environ["EVENT_NAME"] == "pull_request_target":
-              event_path = Path(os.environ["EVENT_PATH"])
-              event = json.loads(event_path.read_text(encoding="utf-8"))
-          else:
-              number = int(os.environ["INPUT_PR_NUMBER"])
-              if number < 1:
-                  raise ValueError("pr_number must be positive")
-              request = Request(
-                  (
-                      "https://api.github.com/repos/"
-                      f"{os.environ['REPOSITORY']}/pulls/{number}"
-                  ),
-                  headers={
-                      "Accept": "application/vnd.github+json",
-                      "Authorization": f"Bearer {os.environ['GH_TOKEN']}",
-                      "X-GitHub-Api-Version": "2022-11-28",
-                  },
-              )
-              with urlopen(request, timeout=30) as response:
-                  pull = json.loads(response.read())
-              event = {"pull_request": pull}
-              event_path = (
-                  Path(os.environ["TEMPORARY_DIRECTORY"])
-                  / "pr-review-event.json"
-              )
-              event_path.write_text(json.dumps(event), encoding="utf-8")
-
-          pull = event["pull_request"]
-          base_sha = pull["base"]["sha"]
-          head_sha = pull["head"]["sha"]
-          if not all(
-              re.fullmatch(r"[0-9a-f]{40}", value)
-              for value in (base_sha, head_sha)
-          ):
-              raise ValueError("pull request revisions must be full SHAs")
-          with Path(os.environ["GITHUB_OUTPUT"]).open("a", encoding="utf-8") as output:
-              output.write(f"base_sha={base_sha}\n")
-              output.write(f"event_path={event_path}\n")
-          PY
-
-      - name: Check out trusted base revision
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-        with:
-          ref: ${{ steps.target.outputs.base_sha }}
-          persist-credentials: false
-
-      - name: Review pull request diff
-        env:
-          GITHUB_TOKEN: ${{ github.token }}
-          LLM_REVIEW_API_KEY: ${{ secrets.LLM_REVIEW_API_KEY }}
-          LLM_REVIEW_BASE_URL: ${{ vars.LLM_REVIEW_BASE_URL }}
-          LLM_REVIEW_MODEL: ${{ vars.LLM_REVIEW_MODEL }}
-          LLM_REVIEW_ID: >-
-            ${{ github.event_name == 'workflow_dispatch' &&
-            'llm-pr-review-canary' || 'llm-pr-review' }}
-          LLM_REVIEW_LABEL: Fast review — diff only
-        run: >-
-          python3 .github/scripts/llm_pr_review.py
-          --event-path "${{ steps.target.outputs.event_path }}"
-          --prompt-path .github/scripts/llm_review_prompt.md
